Implemented solution to day 7, part 1
Implemented solution to part 1 of day 7.
Initially made the stupid decision of parsing the string
of "rules" every time we try to determine what a given bag contains.
My stupidity was made obvious when the solution refused to complete
on my laptop...

Then implemented a Ruleset class which represents the rules
as a HashSet of HashSets, and everything went much smoother!